refactor : 미션3 진행을 위한 패키지 수정
