+++ src/main/java/com/kdt/lecture/customer/CustomerRepository.java
@@ -1,4 +1,4 @@
-package com.kdt.lecture.domain;
+package com.kdt.lecture.customer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
